Updated gitignore
I updated gitignore to included some generated files. In one case, it
was already in gitignore, but with an old file path.

I also added the templates folder to gitignore. Users should manage
their templates outside of the repository. The only templates in the
git repository should be the standard ones that have been tested.

	modified:   .gitignore