Fixing the code so that S8 condensation occurs
Since we are including S4+S4 --> S8 gas, we need to include S8
condensation to the aerosol phase. To do this, I had to remove the
skips8 code in line 331 of the Dochem.f subroutine. I replaced that code
with an IF statement base on 'LS8 .ne. 0'. The code now insures that S8
condensation occurs if liquid (?) sulfur (LS8) is present, and uses the
LS8 variable as a trigger for condensation. This should help the model
converge for high sulfur scenarios.

	modified:   PHOTOCHEM/SUBROUTINES/Dochem.f